Issue #1060: Add sf_get_format_check_failure_reason()
This is a candidate implementation for GitHub issue #1060 -- it adds
a new function to the libsndfile API:

   const char * sf_get_format_check_failure_reason(const SF_INFO *);

This function is logically equivalent to the existing sf_format_check()
function, except that instead of returning SF_TRUE on success and SF_FALSE
on failure, it returns NULL on success and a human-readable explanation
string on failure, so that the nature of the format-incompatibility can
be made more readily-apparent to the developer or the user.

Replaced the existing sf_format_check() implementation with a one-line
wrapper around this function to avoid code-duplication.